Update whitelist.me
Add a domain name to the whitelist

**More Information**
How did you discover your web site or domain was listed here?
Received a report from a user of MalwareBytes and viewed a report in virustotal

**Have you requested removal from other sources?**
Requested from Webroot and CRDF.

**Additional context**
A trial user was able to upload a file that was then used in a phishing attempt some months ago. This was detected and removed by us within 24 hours, and actions taken to prevent this from happening again.